--- Document Apis Information.txt
@@ -1,18 +1,0 @@
-APIS
-
-/api/v1/users
-				GET -> obtener todos los usuarios
-
-
-/api/v1/user
-				GET : id -----> required   ---> hay un bug, cuando no existe el id rompe todo
-				PUT : id -----> required   ---> hay un bug, cuando no existe el id rompe todo, funciona como post tambien OJO
-				DELETE : id --> required   ---> hay un bug, cuando no existe el id rompe todo
-
-/api/v1/areas
-				GET -> obtener todos las areas
-
-/api/v1/area
-				GET : id ----> required   --->  hay un bug, cuando no existe el id rompe todo
-				PUT : id ----> required   --->  hay un bug, cuando no existe el id rompe todo, funciona como post tambien OJO
-				DELETE: id --> required   --->  hay un bug, cuando no existe el id rompe todo
