--- data/postMethod.js
@@ -1,32 +1,0 @@
-const connection = require('./connection');
-let ObjectId = require('mongodb').ObjectId;
-const bcrypt = require('bcryptjs');
-
-async function addUser(user) {
-    const mongoClient = await connection.getConnection();
-    user.password = bcrypt.hashSync(user.password, 8);
-    const result = await mongoClient.db('ReservasPuesto')
-        .collection('users')
-        .insertOne(user);
-    return result;
-}
-
-async function addArea(area) {
-    const mongoClient = await connection.getConnection();
-    const result = await mongoClient.db('ReservasPuesto')
-        .collection('areas')
-        .insertOne(area);
-    return result;
-}
-
-async function addReservation(date) {
-    let reserva = { date: date, usersId: [] }
-    const mongoClient = await connection.getConnection();
-    const result = await mongoClient.db('ReservasPuesto')
-        .collection('reservas')
-        .insertOne(reserva);
-    return result;
-}
-
-
-module.exports = { addUser, addArea, addReservation };
