+++ middleware/authenticator.js
@@ -2,13 +2,15 @@ const express = require('express');
 const router = express.Router();
 
 router.use('/index', (req, res, next) => {
-  if (!req.session.user) {
+  /* if (!req.session.user) {
     req.flash('message', '❌ Debe iniciar sesión ❌')
     res.redirect("/")
   }
   else {
     next()
-  }
+  } */
+  console.log('api');
+  next();
 })
 
 module.exports = router;
