+++ package.json
@@ -23,8 +23,6 @@
     "chalk": "^3.0.0",
     "ejs": "^3.1.6",
     "express": "^4.17.1",
-    "express-flash": "^0.0.2",
-    "express-session": "^1.17.2",
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.6.7",
     "nodemon": "^2.0.7"
