--- routers/apis.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const getDataMethod = require('../data/getMethod');
-const putDataMethod = require('../data/putMethod');
-const deleteDataMethod = require('../data/deleteMethod');
-const postDataMethod = require('../data/postMethod');
-
-
-//autenticacion
