+++ routers/area.js
@@ -1,12 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const getDataMethod = require('../data/getMethod');
-const putDataMethod = require('../data/putMethod');
-const deleteDataMethod = require('../data/deleteMethod');
-const postDataMethod = require('../data/postMethod');
+const area = require('../data/area');
 
 router.get('/api/v1/areas', async (req, res) => {
-    const areas = await getDataMethod.getAreas();
+    const areas = await area.getAreas();
     res.json(areas);
 })
 
@@ -17,23 +14,23 @@ router.get('/api/v1/area', async (req, res) => {
         res.send('id requerido')
     } else {
         if (idQuery) {
-            const area = await getDataMethod.getArea(idQuery);
-            res.json(area);
+            const searchArea = await area.getArea(idQuery);
+            res.json(searchArea);
         }
         if (idBody) {
-            const area = await getDataMethod.getArea(idBody);
-            res.json(area);
+            const searchArea = await area.getArea(idBody);
+            res.json(searchArea);
         }
     }
 });
 
 router.post('/api/v1/area', async (req, res) => {
-    let area = req.body.area;
-    if (!area) {
+    let newArea = req.body.area;
+    if (!newArea) {
         res.send('parametro requerido en POST "area", los parametros deben enviarse por el body')
     } else {
-        area = await postDataMethod.addArea(req.body);
-        res.json(area.ops)
+        newArea = await area.addArea(req.body);
+        res.json(newArea.ops)
     }
 });
 
@@ -47,7 +44,7 @@ router.put('/api/v1/area', async (req, res) => {
     } else {
         if (idBody) {
             let userBody = req.body;
-            userBody = await putDataMethod.updateArea(userBody, idBody);
+            userBody = await area.updateArea(userBody, idBody);
             if (!userBody) {
                 res.send('No se pudo procesar la solicitud, verificar los valores enviados')
             } else {
@@ -57,7 +54,7 @@ router.put('/api/v1/area', async (req, res) => {
 
         if (idQuery) {
             let userQuery = req.query;
-            userQuery = await putDataMethod.updateArea(userQuery, idQuery);
+            userQuery = await area.updateArea(userQuery, idQuery);
             if (!userQuery) {
                 res.send('No se pudo procesar la solicitud, verificar los valores enviados')
             } else {
@@ -76,11 +73,11 @@ router.delete('/api/v1/area', async (req, res) => {
         res.send('ID requerido')
     } else {
         if (idQuery) {
-            await deleteDataMethod.deleteArea(idQuery);
+            await area.deleteArea(idQuery);
             res.send('Usuario Eliminado')
         }
         if (idBody) {
-            await deleteDataMethod.deleteArea(idBody);
+            await area.deleteArea(idBody);
             res.send('Usuario Eliminado')
         }
     }
