--- routers/calendar.js
@@ -1,5 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const getDataMethod = require('../data/getMethod');
-const postDataMethod = require('../data/postMethod');
-
