--- routers/logout.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.get('', (req, res) => {
-    req.session.destroy();
-    res.redirect('/')
-  })
-
-module.exports = router;
