+++ routers/reservation.js
@@ -1,12 +1,10 @@
 const express = require('express');
 const router = express.Router();
-const getDataMethod = require('../data/getMethod');
-const putDataMethod = require('../data/putMethod');
-const deleteDataMethod = require('../data/deleteMethod');
-const postDataMethod = require('../data/postMethod');
+const user = require('../data/user')
+const reservation = require('../data/reservation');
 
 router.get('/api/v1/reservation', async (req, res) => {
-    const users = await getDataMethod.getUsers();
+    const users = await user.getUsers();
     if (users.length == 0) {
         res.status(204);
     }
@@ -16,11 +14,11 @@ router.get('/api/v1/reservation', async (req, res) => {
 
 router.get('/api/v1/reservation/:id', async (req, res) => {
     let { id } = req.params;
-    let user = await getDataMethod.getUser(id);
-    if (!user) {
+    let searchUser = await user.getUser(id);
+    if (!searchUser) {
         return res.status(404).json({ "error": "usuario no encontrado" });
     }
-    return res.json(user);
+    return res.json(searchUser);
 })
 
 router.post('/api/v1/reservation', async (req, res) => {
@@ -29,25 +27,25 @@ router.post('/api/v1/reservation', async (req, res) => {
         return res.status(400)
             .json({ "error": "parametros requeridos en POST 'date' 'userId', los parametros deben enviarse por el body" });
     } else {
-        let user = await getDataMethod.getUser(userId);
-        if (!user) {
+        let searchUser = await user.getUser(userId);
+        if (!searchUser) {
             return res.status(400)
                 .json({ "error": "Usuario no encontrado" });
         }
-        let reserva = await getDataMethod.getReservationByDate(date)
+        let reserva = await reservation.getReservationByDate(date)
         console.log("reserva", reserva)
         console.log("#########")
         let reservaId = "";
         if (!reserva) {
-            reserva = await postDataMethod.addReservation(date)
+            reserva = await reservation.addReservation(date)
             console.log("new reserva", reserva)
             reservaId = reserva.ops[0]._id;
         } else {
             reservaId = reserva._id;
         }
-        putDataMethod.addUserToReservation(userId, reservaId);
+        user.addUserToReservation(userId, reservaId);
 
-        return res.status(201).json({ "success": true, "usuario": user });
+        return res.status(201).json({ "success": true, "usuario": searchUser });
     }
 });
 
@@ -56,7 +54,7 @@ router.put('/api/v1/reservation/:id', async (req, res) => {
     if (!id) {
         return res.status(400).json({ "error": "el parametro _id es requerido" });
     }
-    let updatedUser = await putDataMethod.updateUser(req.body, id);
+    let updatedUser = await user.updateUser(req.body, id);
     console.log('updatedUser:', updatedUser);
     if (!updatedUser) {
         return res.status(404).json({ "error": "el usuario no existe" });
@@ -71,12 +69,12 @@ router.delete('/api/v1/reservation/:id', async (req, res) => {
     if (!id) {
         return res.status(400).json({ "error": "el parametro _id es requerido" });
     }
-    let user = await getDataMethod.getUser(id);
-    if (!user) {
+    let searchUser = await user.getUser(id);
+    if (!searchUser) {
         return res.status(404).json({ "error": "usuario no encontrado" });
     }
     await deleteDataMethod.deleteUser(id);
-    return res.json({ "success": true, "deletedUser": user });
+    return res.json({ "success": true, "deletedUser": searchUser });
 });
