+++ routers/user.js
@@ -1,45 +1,44 @@
 const express = require('express');
 const router = express.Router();
-const getDataMethod = require('../data/getMethod');
-const putDataMethod = require('../data/putMethod');
-const deleteDataMethod = require('../data/deleteMethod');
-const postDataMethod = require('../data/postMethod');
+const user = require('../data/user');
+
+const profile = require('../data/profile');
+const reservation = require('../data/reservation');
 
 router.get('/api/v1/users', async (req, res) => {
-    const users = await getDataMethod.getUsers();
+    const users = await user.getUsers();
     if (users.length == 0) {
         res.status(204);
     }
     return res.json({ "count": users.length, users });
-
 })
 
-router.get('/api/v1/users/:id', async (req, res) => {
+router.get('/api/v1/user/:id', async (req, res) => {
     let { id } = req.params;
-    let user = await getDataMethod.getUser(id);
-    if (!user) {
+    let searchUser = await user.getUser(id);
+    if (!searchUser) {
         return res.status(404).json({ "error": "usuario no encontrado" });
     }
-    return res.json(user);
+    return res.json(searchUser);
 })
 
-router.post('/api/v1/users', async (req, res) => {
+router.post('/api/v1/user', async (req, res) => {
     let { email, password, profile, area } = req.body;
     if (!email || !password || !profile || !area) {
         return res.status(400)
             .json({ "error": "parametros requeridos en POST 'email' 'contrasena' 'perfil' 'area', los parametros deben enviarse por el body" });
     } else {
-        let user = await postDataMethod.addUser(req.body);
-        return res.status(201).json({ "success": true, "usuario": user.ops[0] });
+        let newUser = await user.addUser(req.body);
+        return res.status(201).json({ "success": true, "usuario": newUser.ops[0] });
     }
 });
 
-router.put('/api/v1/users/:id', async (req, res) => {
+router.put('/api/v1/user/:id', async (req, res) => {
     let { id } = req.params;
     if (!id) {
         return res.status(400).json({ "error": "el parametro _id es requerido" });
     }
-    let updatedUser = await putDataMethod.updateUser(req.body, id);
+    let updatedUser = await user.updateUser(req.body, id);
     console.log('updatedUser:', updatedUser);
     if (!updatedUser) {
         return res.status(404).json({ "error": "el usuario no existe" });
@@ -54,12 +53,12 @@ router.delete('/api/v1/users/:id', async (req, res) => {
     if (!id) {
         return res.status(400).json({ "error": "el parametro _id es requerido" });
     }
-    let user = await getDataMethod.getUser(id);
-    if (!user) {
+    let deleteUser = await user.getUser(id);
+    if (!deleteUser) {
         return res.status(404).json({ "error": "usuario no encontrado" });
     }
-    await deleteDataMethod.deleteUser(id);
-    return res.json({ "success": true, "deletedUser": user });
+    await user.deleteUser(id);
+    return res.json({ "success": true, "deletedUser": deleteUser });
 });
 
 module.exports = router;
