--- views/abmArea.ejs
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Reserva tu lugar</title>
-    <link rel="stylesheet" href="../stylesheets/estilos.css">
-    <link rel="stylesheet" href="../stylesheets/mini-default.min.css">
-</head>
-
-<body>
-    <%- include ('./partials/back') %>
-        <div class="container">
-            <h1>Usuarios</h1>
-            <hr>
-            <div class="row">
-                <div class="col-sm-12">
-                    <a class="button" href="/admin/createArea">Agregar Area</a>
-                    <table>
-                        <thead>
-                            <tr>
-                                <th>Area</th>
-                                <th>Opciones</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <% area.forEach((a)=>{ %>
-                                <tr>
-                                    <td>
-                                        <%=a.area%>
-                                    </td>
-                                    <td>
-                                        <a href="/admin/editArea/<%=a._id%>"><span class="icon-edit"></span>
-                                            Editar
-                                        </a>
-                                        |
-                                        <a href="/admin/delete_area_process/<%=a._id%>"
-                                            onclick="return confirm('¿Está seguro de eliminar: <%=a.area%>?')"><span
-                                                class="icon-alert"></span>
-                                            Eliminar
-                                        </a>
-                                    </td>
-                                </tr>
-                                <%}) %>
-                        </tbody>
-                    </table>
-                    <hr>
-                </div>
-            </div>
-        </div>
-        <%- include ('./partials/piePagina') %>
-        <script src="javascripts/script.js"></script>
-</body>
-
-</html>
