--- views/abmUser.ejs
@@ -1,69 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Reserva tu lugar</title>
-    <link rel="stylesheet" href="../stylesheets/estilos.css">
-    <link rel="stylesheet" href="../stylesheets/mini-default.min.css">
-</head>
-
-<body>
-    <%- include ('./partials/back') %>
-        <div class="container">
-            <% if(error && error.length > 0) { %>
-                <h3><span class="icon-info"></span><%=error%></h3>
-            <%}%>
-
-            <h1>Usuarios</h1>
-            <hr>
-            <div class="row">
-                <div class="col-sm-12">
-                    <a class="button" href="/admin/userRegister">Agregar Usuario</a>
-                    <table>
-                        <thead>
-                            <tr>
-                                <th>Email</th>
-                                <th>Perfil</th>
-                                <th>Area</th>
-                                <th>Opciones</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <% user.forEach((user)=>{ %>
-                                <tr>
-                                    <td>
-                                        <%=user.email%>
-                                    </td>
-                                    <td>
-                                        <%=user.profile%>
-                                    </td>
-                                    <td>
-                                        <%=user.area%>
-                                    </td>
-                                    <td>
-                                        <a href="/admin/edit/<%=user._id%>"><span class="icon-edit"></span>
-                                            Editar
-                                        </a>
-                                        |
-                                        <a href="/admin/delete_process/<%=user._id%>"
-                                            onclick="return confirm('¿Está seguro de eliminar: <%=user.email%>?')"><span
-                                                class="icon-alert"></span>
-                                            Eliminar
-                                        </a>
-                                    </td>
-                                </tr>
-                                <%}) %>
-                        </tbody>
-                    </table>
-                    <hr>
-                </div>
-            </div>
-        </div>
-        <%- include ('./partials/piePagina') %>
-        <script src="javascripts/script.js"></script>
-</body>
-
-</html>
