--- views/createArea.ejs
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Reserva tu lugar</title>
-    <link rel="stylesheet" href="../stylesheets/estilos.css">
-    <link rel="stylesheet" href="../stylesheets/mini-default.min.css">
-</head>
-
-<body>
-    <%- include ('./partials/encabezadoPrivado') %>
-
-        <div class="container">
-            <h1>Registrar Area</h1>
-            <p>Ingresar Datos</p>
-
-            <div class="row">
-                <div class="col-sm-12">
-
-                    <form action="/admin/area_register_process" method="post">
-                        <fieldset>
-                            <legend>Nueva Area</legend>
-                            <p>
-                                <label for="newArea">Nombre del Area</label> <br>
-                                <input type="text" name="area" id="area" required="true">
-                            </p>
-                            <p>
-                                <input class="primary" type="submit" value="Continuar">
-                            </p>
-                        </fieldset>
-                    </form>
-                </div>
-            </div>
-        </div>
-
-        <%- include ('./partials/piePagina') %>
-            <script src="../javascripts/script.js"></script>
-</body>
-
-</html>
