--- views/editArea.ejs
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>EditarArea</title>
-  <link rel="stylesheet" href="/stylesheets/estilos.css">
-  <link rel="stylesheet" href="/stylesheets/mini-default.min.css">
-</head>
-
-<body>
-  <%- include ('./partials/back') %>
-
-    <div class="container">
-      <h1>Admin - Editar Area</h1>
-      <hr>
-      <div class="row">
-        <div class="col-sm-12">
-
-          <form action="/admin/edit_area_process" method="POST">
-            <input type="hidden" name="_id" value="<%=area_id._id%>">
-            <fieldset>
-              <legend>Datos del Area</legend>
-              <p>
-                <label for="name">Area</label>
-                <br>
-                <input type="text" name="area" id="" value="<%=area_id.area%>">
-              </p>
-              <p>
-                <input class="primary" type="submit" value="Continuar" />
-              </p>
-            </fieldset>
-          </form>
-        </div>
-      </div>
-    </div>
-
-    <%- include ('./partials/piePagina') %>
-    <script src="/javascripts/script.js"></script>
-</body>
-
-</html>
