--- views/login.ejs
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-
-<head>
-    <title>Inicio de Sesion</title>
-
-    <!--JQUERY-->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-
-    <!-- FRAMEWORK BOOTSTRAP para el estilo de la pagina-->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-
-    <!-- Los iconos tipo Solid de Fontawesome-->
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.8/css/solid.css">
-    <script src="https://use.fontawesome.com/releases/v5.0.7/js/all.js"></script>
-
-    <!-- Nuestro css-->
-    <link rel="stylesheet" type="text/css" href="stylesheets/styles.css" th:href="@{/css/styles.css}">
-
-</head>
-
-<body>
-    <div class="modal-dialog text-center">
-        <div class="col-sm-8 main-section">
-            <div class="modal-content">
-                <div class="col-12 user-img">
-                    <img src="images/user.png" th:src="@{/img/user.png}" />
-                </div>
-                <form class="col-12" action="/login-process" method="post">
-                    <div class="form-group" id="user-email">
-                        <input type="email" class="form-control" placeholder="Email 📧" name="email" id="" required="true">
-                    </div>
-                    <div class="form-group" id="contrasena-group">
-                        <input type="password" class="form-control" placeholder="Contraseña 🔐" name="password"
-                            required="true" />
-                    </div>
-                    <button type="submit" class="btn btn-primary"><i class="fas fa-sign-in-alt"></i> Ingresar </button>
-                </form>
-                
-                <% if(message && message.length > 0){ %>
-                    <div class="alert alert-danger" role="alert">
-                        <span>
-                            <%=message%>
-                        </span>
-                    </div>
-                <% } %>
-            </div>
-        </div>
-    </div>
-</body>
-
-</html>
