--- views/partials/encabezadoPrivado.ejs
@@ -1,10 +1,0 @@
-<header class="sticky">
-    <div class="enlaces-registro">
-        <span class="icon-user"></span>
-        <%=user.email%> | <%=user.profile%>
-                <a href="/logout">Cerrar Sesion</a>
-    </div>
-    <a href="/index/reserve-place" class="button"><span class="icon-credit"></span>Reservar</a>
-    <a href="/admin/abm-user" class="button"><span class="icon-credit"></span>ABM de usuario</a>
-    <a href="/admin/abm-area" class="button"><span class="icon-credit"></span>ABM de Areas</a>
-</header>
