--- views/partials/encabezadoUsuario.ejs
@@ -1,9 +1,0 @@
-<header class="sticky">
-  <div class="enlaces-registro">
-    <span class="icon-user"></span>
-    <%=user.email%> | <%=user.profile%>
-        <a href="/logout">Cerrar Sesion</a>
-  </div>
-  <a href="#" class="logo">Reserva Tu Lugar</a>
-  <a href="/index/reserve-place" class="button"><span class="icon-credit"></span>Reservar Lugar</a>
-</header>
