--- views/partials/piePagina.ejs
@@ -1,3 +1,0 @@
-<footer>
-    <p>Reserva tu lugar - Vive la experiencia - Derechos Reservados</p>
-</footer>
