--- views/registerAbmUser.ejs
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>RegistroAbmUsuario</title>
-    <link rel="stylesheet" href="../stylesheets/estilos.css">
-    <link rel="stylesheet" href="../stylesheets/mini-default.min.css">
-</head>
-
-<body>
-    <%- include ('./partials/encabezadoPrivado') %>
-
-        <div class="container">
-            <h1>Registrar nuevo usuario</h1>
-            <p>Ingresar Datos</p>
-
-            <div class="row">
-                <div class="col-sm-12">
-                            <form action="/admin/user_register_process" method="post">
-                                <fieldset>
-                                    <legend>Nuevo Usuario</legend>
-                                    <p>
-                                        <label for="email">Email</label> <br>
-                                        <input type="email" name="email" id="email" required="true">
-                                    </p>
-                                    <p>
-                                        <label for="password">Contraseña</label><br>
-                                        <input type="password" name="password" id="password" required="true">
-                                    </p>
-                                    <p>
-                                        <label for="profile">Perfil</label>
-                                        
-                                        <select class="form-select" name="profile" aria-label="Default select example" required="true">
-                                            <option value="" selected>Seleccione</option>
-                                            <% profiles.forEach((profile)=>{ %>
-                                            <option value="<%=profile.profile%>"><%=profile.profile%></option>
-                                            <%}) %>
-                                        </select>
-                                    </p>
-                                    <p>
-                                        <label for="area">Area</label>
-                                        <select class="form-select" name="area" aria-label="Default select example" required="true">
-                                            <option value="" selected>Seleccione</option>
-                                            <% areas.forEach((area)=>{ %>
-                                            <option value="<%=area.area%>"><%=area.area%></option>
-                                            <%}) %>
-                                        </select>
-                                    </p>
-                                    <p>
-                                        <input class="primary" type="submit" value="Continuar">
-                                    </p>
-                                </fieldset>
-                            </form>
-                </div>
-            </div>
-        </div>
-
-        <%- include ('./partials/piePagina') %>
-            <script src="../javascripts/script.js"></script>
-</body>
-
-</html>
